+++ app/src/test/java/woowacourse/shopping/TestFixtures.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.viewmodel
+package woowacourse.shopping
 
 val imageUrl = "https://www.naver.com/"
 val title = "올리브"
